+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [0.9.5]
 * Fixed compatibility with Emscripten 3.1.44
+* Improved support for `Pointer<Utf8>`
 
 ## [0.9.4]
 * Fixed analyzer warnings