+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [0.9.5]
+* Fixed compatibility with Emscripten 3.1.44
+
 ## [0.9.4]
 * Fixed analyzer warnings